Moved to docker compose watch instead of mounting volumes
Improves performance as the local node_modules can be copied in, but we'll still rebuild the container when the yarn.lock file is changed.

Had to move to nodemon as Docker sync moves in a new file with the same name, causing `node --watch` to only see the first update.
https://forums.docker.com/t/compose-watch-with-node-watch-only-reloads-first-time/138072/4